Update book1_TobieBarry.txt
Zombie apocalypse incoming! Plus space aliens? And elves are real. CoOl. And if you eat a lot of protein after geTting hit in the head, it's easier. Sigura had found that out through first hand experience. But... it had to be vegan, or the nanomachines wouldn't work. Well, they would, but they would eat meat, which were made out of.